+++ mapbox_baselayer/migrations/0006_alter_baselayertile_id_alter_mapbaselayer_id_and_more.py
@@ -15,7 +15,7 @@ class Migration(migrations.Migration):
             name="map_box_url",
             field=models.CharField(
                 blank=True,
-                help_text="Mapbox or tilejson URL, starting ",
+                help_text="Mapbox or tilejson URL",
                 max_length=255,
             ),
         ),