+++ mapbox_baselayer/urls.py
@@ -5,5 +5,6 @@ from mapbox_baselayer import views
 app_name = 'mapbox_baselayer'
 
 urlpatterns = [
-    path('mapbox-baselayers/<int:pk>/tilejson/', views.MapboxBaseLayerJsonDetailView.as_view(), name='tilejson')
+    path('mapbox-baselayers/<int:pk>/tilejson/', views.MapboxBaseLayerJsonDetailView.as_view(), name='tilejson'),
+    path('mapbox-baselayers/', views.MapLayerListView.as_view(), name='baselayer-list'),
 ]