+++ test_mapbox_baselayer/settings.py
@@ -38,8 +38,11 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'mapbox_baselayer',
+    'test_mapbox_baselayer.test_app',
 ]
-
+LOCALE_DIRS = [
+    os.path.join(BASE_DIR, 'locale'),
+]
 MIDDLEWARE = [
     'django.middleware.security.SecurityMiddleware',
     'django.middleware.locale.LocaleMiddleware',